+++ HISTORY.rst
@@ -22,6 +22,7 @@ Unreleased
 * ALCYN-1417: Fix ESB connector following addition of customer hierarchy
 * ALCYN-1432: Create the tax 100 outside EU
 * ALCYN-1433: Archive outside EU taxes (21, 12%, 6% and 0%)
+* ALCYN-1434: Update tax "TVA à l'entrée 0% Hors EU EXTRACOM - Approvisionn. et marchandises"
 
 **Bugfixes**